@@ -1,6 +1,6 @@
 {
   "name": "custom-form-app",
-  "version": "0.0.0",
+  "version": "1.1.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -14,40 +14,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.15",
-    "@angular/cdk": "~7.3.7",
-    "@angular/common": "~7.2.15",
-    "@angular/compiler": "~7.2.15",
-    "@angular/core": "~7.2.15",
-    "@angular/elements": "^7.2.15",
-    "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "~7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/material-moment-adapter": "^7.3.7",
-    "@angular/platform-browser": "~7.2.15",
-    "@angular/platform-browser-dynamic": "~7.2.15",
-    "@angular/router": "~7.2.15",
+    "@angular/animations": "~8.1.0",
+    "@angular/cdk": "~8.0.2",
+    "@angular/common": "~8.1.0",
+    "@angular/compiler": "~8.1.0",
+    "@angular/core": "~8.1.0",
+    "@angular/elements": "^8.1.0",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@angular/forms": "~8.1.0",
+    "@angular/material": "^8.0.2",
+    "@angular/material-moment-adapter": "^8.0.2",
+    "@angular/platform-browser": "~8.1.0",
+    "@angular/platform-browser-dynamic": "~8.1.0",
+    "@angular/router": "~8.1.0",
     "moment": "^2.24.0",
     "concat": "^1.0.3",
-    "core-js": "^2.5.4",
+    "core-js": "^3.1.4",
     "document-register-element": "^1.7.2",
     "fs-extra": "^7.0.1",
     "hammerjs": "^2.0.8",
-    "rxjs": "~6.3.3",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "rxjs": "~6.5.2",
+    "tslib": "^1.9.3",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^7.4.3",
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.9",
-    "@angular/compiler-cli": "~7.2.15",
-    "@angular/language-service": "~7.2.15",
+    "@angular-builders/custom-webpack": "^8.0.2",
+    "@angular-devkit/build-angular": "~0.801.0",
+    "@angular/cli": "~8.1.0",
+    "@angular/compiler-cli": "~8.1.0",
+    "@angular/language-service": "~8.1.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
+    "@types/node": "~8.10.39",
     "@webcomponents/webcomponentsjs": "^2.2.10",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "~5.1.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.0.0",
@@ -56,8 +56,8 @@
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
+    "ts-node": "~7.0.1",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "typescript": "~3.4.5"
   }
 }
